# Less hops in operator lifting.