--- a/Sources/SignalProducer.swift
+++ b/Sources/SignalProducer.swift
@@ -2,26 +2,28 @@ import Dispatch
 import Foundation
 import Result
 
-/// A `SignalProducer` produces a `Signal` with an instance of repeatable work which
-/// yields events to the `Signal`, every time it is started with an observer.
+/// A SignalProducer creates Signals that can produce values of type `Value`
+/// and/or fail with errors of type `Error`. If no failure should be possible,
+/// `NoError` can be specified for `Error`.
 ///
-/// Even if multiple instances of work from the same `SignalProducer` are executing
-/// concurrently, produced `Signal`s would see only events from its own instance of work,
-/// but not from the others.
+/// SignalProducers can be used to represent operations or tasks, like network
+/// requests, where each invocation of `start()` will create a new underlying
+/// operation. This ensures that consumers will receive the results, versus a
+/// plain Signal, where the results might be sent before any observers are
+/// attached.
 ///
-/// The _work_ performed by a `SignalProducer` is also known as the _post-creation side
-/// effect_.
-///
-/// `SignalProducer`s can be used to represent repeatable operations or tasks — like
-/// network requests — that are customizable, lazy and/or on demand.
+/// Because of the behavior of `start()`, different Signals created from the
+/// producer may see a different version of Events. The Events may arrive in a
+/// different order between Signals, or the stream might be completely
+/// different!
 public struct SignalProducer<Value, Error: Swift.Error> {
 	public typealias ProducedSignal = Signal<Value, Error>
 
 	/// Wraps a closure which, when invoked, produces a new instance of `Signal`, a
-	/// customized `didCreate` post-creation side effect for the `Signal` and a disposable
-	/// to interrupt the produced `Signal`.
+	/// customized `observerDidSetup` post-creation side effect for the `Signal` and a
+	/// disposable to interrupt the produced `Signal`.
 	///
-	/// Unlike the safe `startWithSignal(_:)` API, `Builder` shifts the responsibility of
+	/// Unlike the safe `startWithSignal(_:)` API, `builder` shifts the responsibility of
 	/// invoking the post-creation side effect to the caller, while it takes from the
 	/// caller the responsibility of the `Signal` creation.
 	///
@@ -30,13 +32,14 @@ public struct SignalProducer<Value, Error: Swift.Error> {
 	/// `startWithSignal(_:)`, regardless of the fact that lifted operators can rely on
 	/// the upstreams for producer interruption.
 	///
-	/// If the caller is a `Builder`, it must invoke the produced `didCreate` before
-	/// performing any of its own post-creation side effect.
-	fileprivate struct Builder {
-		fileprivate let make: () -> (signal: Signal<Value, Error>, didCreate: () -> Void, interruptHandle: Disposable)
+	/// `observerDidSetup` must be invoked before any other post-creation side effect.
+	fileprivate struct Instance {
+		let producedSignal: Signal<Value, Error>
+		let observerDidSetup: () -> Void
+		let interruptHandle: Disposable
 	}
 
-	fileprivate let builder: Builder
+	fileprivate let builder: () -> Instance
 
 	/// Initializes a `SignalProducer` that will emit the same events as the
 	/// given signal.
@@ -68,14 +71,14 @@ public struct SignalProducer<Value, Error: Swift.Error> {
 	/// - parameters:
 	///   - startHandler: A closure that accepts observer and a disposable.
 	public init(_ startHandler: @escaping (Signal<Value, Error>.Observer, Lifetime) -> Void) {
-		self.init(Builder {
+		self.init { _ -> Instance in
 			let disposable = CompositeDisposable()
 			let (signal, observer) = Signal<Value, Error>.pipe(disposable: disposable)
-			let didCreate = { startHandler(observer, Lifetime(disposable)) }
+			let observerDidSetup = { startHandler(observer, Lifetime(disposable)) }
 			let interruptHandle = AnyDisposable(observer.sendInterrupted)
 
-			return (signal, didCreate, interruptHandle)
-		})
+			return Instance(producedSignal: signal, observerDidSetup: observerDidSetup, interruptHandle: interruptHandle)
+		}
 	}
 
 	/// Create a SignalProducer that will invoke the given factory once for each
@@ -83,7 +86,7 @@ public struct SignalProducer<Value, Error: Swift.Error> {
 	///
 	/// - parameters:
 	///   - builder: A builder that is used by `startWithSignal` to create new `Signal`s.
-	fileprivate init(_ builder: Builder) {
+	fileprivate init(_ builder: @escaping () -> Instance) {
 		self.builder = builder
 	}
 
@@ -220,10 +223,10 @@ public struct SignalProducer<Value, Error: Swift.Error> {
 	///            `Signal` commences. Both the produced `Signal` and an interrupt handle
 	///            of the signal would be passed to the closure.
 	public func startWithSignal(_ setup: (_ signal: Signal<Value, Error>, _ interruptHandle: Disposable) -> Void) {
-		let (signal, didCreate, interruptHandle) = builder.make()
-		setup(signal, interruptHandle)
-		guard !interruptHandle.isDisposed else { return }
-		didCreate()
+		let instance = builder()
+		setup(instance.producedSignal, instance.interruptHandle)
+		guard !instance.interruptHandle.isDisposed else { return }
+		instance.observerDidSetup()
 	}
 }
 
@@ -404,12 +407,14 @@ extension SignalProducer {
 	/// - returns: A signal producer that applies signal's operator to every
 	///            created signal.
 	public func lift<U, F>(_ transform: @escaping (Signal<Value, Error>) -> Signal<U, F>) -> SignalProducer<U, F> {
-		return SignalProducer<U, F>(SignalProducer<U, F>.Builder {
+		return SignalProducer<U, F> { _ -> SignalProducer<U, F>.Instance in
 			// Transform the `Signal`, and pass through the `didCreate` side effect and
 			// the interruptHandle.
-			let (signal, didCreate, interruptHandle) = self.producer.builder.make()
-			return (signal: transform(signal), didCreate: didCreate, interruptHandle: interruptHandle)
-		})
+			let instance = self.producer.builder()
+			return SignalProducer<U, F>.Instance(producedSignal: transform(instance.producedSignal),
+			                                     observerDidSetup: instance.observerDidSetup,
+			                                     interruptHandle: instance.interruptHandle)
+		}
 	}
 
 	/// Lift a binary Signal operator to operate upon SignalProducers.
@@ -438,22 +443,21 @@ extension SignalProducer {
 
 	private func lift<U, F, V, G>(leftFirst: Bool, _ transform: @escaping (Signal<Value, Error>) -> (Signal<U, F>) -> Signal<V, G>) -> (SignalProducer<U, F>) -> SignalProducer<V, G> {
 		return { otherProducer in
-			return SignalProducer<V, G>(SignalProducer<V, G>.Builder {
-				let (left, didCreateLeft, leftInterrupter) = self.producer.builder.make()
-				let (right, didCreateRight, rightInterrupter) = otherProducer.builder.make()
+			return SignalProducer<V, G> { _ -> SignalProducer<V, G>.Instance in
+				let left = self.producer.builder()
+				let right = otherProducer.builder()
 
-				return (signal: transform(left)(right),
-				        didCreate: {
-							if leftFirst {
-								didCreateLeft()
-								didCreateRight()
-							} else {
-								didCreateRight()
-								didCreateLeft()
-							}
-				        },
-				        interruptHandle: CompositeDisposable([leftInterrupter, rightInterrupter]))
-			})
+				return .init(producedSignal: transform(left.producedSignal)(right.producedSignal),
+				             observerDidSetup: {
+								if leftFirst {
+									left.observerDidSetup()
+									right.observerDidSetup()
+								} else {
+									right.observerDidSetup()
+									left.observerDidSetup()
+								}},
+				             interruptHandle: CompositeDisposable([left.interruptHandle, right.interruptHandle]))
+			}
 		}
 	}
 
@@ -1592,18 +1596,20 @@ extension SignalProducer {
 		disposed: (() -> Void)? = nil,
 		value: ((Value) -> Void)? = nil
 	) -> SignalProducer<Value, Error> {
-		return SignalProducer(Builder {
-			let (signal, start, interruptHandle) = self.producer.builder.make()
-			return (signal: signal.on(event: event,
-			                          failed: failed,
-			                          completed: completed,
-			                          interrupted: interrupted,
-			                          terminated: terminated,
-			                          disposed: disposed,
-			                          value: value),
-			        didCreate: { starting?(); start(); started?() },
-			        interruptHandle: interruptHandle)
-		})
+		return SignalProducer { _ -> Instance in
+			let instance = self.producer.builder()
+			let signal = instance.producedSignal.on(event: event,
+			                                        failed: failed,
+			                                        completed: completed,
+			                                        interrupted: interrupted,
+			                                        terminated: terminated,
+			                                        disposed: disposed,
+			                                        value: value)
+
+			return Instance(producedSignal: signal,
+			                observerDidSetup: { starting?(); instance.observerDidSetup(); started?() },
+			                interruptHandle: instance.interruptHandle)
+		}
 	}
 
 	/// Start the returned producer on the given `Scheduler`.