#### TL;DR
Closure magic that benefits the heavily used `SignalProducer` ~~and `Property`~~. 🍨 

#### Details
Lifted operators:
1. do not interrupt the producer but only pass through `interrupted` events; and
2. do not create extra resources when lifted.

This means for these operators, the extra levels of indirection of `Signal` created by the nested `startWithSignal` calls is just burning CPU times for no purpose...

The PR breaks the start handler signature (internally) to separate the creation of the `Signal` and the starting side effect.

TL;DR: lifted operators now create ONE signal only, not two, when started. :tada: 

`lift`, `liftLeft`, `liftRight`~~, `SignalProducer(signal:)`~~ and `on` are reimplemented using this new `SignalProducer` format.

That's said it is kept internal for the moment, since it is prone to error, e.g. not calling the upstream's `start`.

#### Note
Composed properties would not benefit from this, since there is an implicit `replayLazily` sitting between lifted operators, which prevents the start handler from coalescing. #144 would help though.

#### Benchmark
[Benchmark Source](https://gist.github.com/andersio/7116ff83e8b5865829b284c288bc0270)
i5 4258U, macOS 10.12.1, Xcode 8.1, `-Owholemodule`

`map` a producer 32 times, and start the producer 1000 times.

Old `lift`: 0.701 sec (3%)
New `lift`: 0.318 sec (6%)

`map` a producer 8 times, and start the producer 4000 times.

Old `lift`: 0.749 sec (2%)
New `lift`: 0.396 sec (4%)