@andersio 2016-12-03
It seems intermediate signals do not get `interrupted` events but are disposed of directly. Fixing.