@andersio 2016-12-07
Let's define some terminologies.

1. Starting a `SignalProducer` graph would create a `Signal` graph.
2. The started `SignalProducer` and the produced `Signal` are considered the root node of the respective graphs.

Note that the order is the reverse of what we usually called upstream and downstream producers. The ultimate upstreams are the leaves in the graph, and the ultimate downstream (that is being explicitly started by the user) is the root of the graph.

#### Interruptible Producer
The graph is constructed with a **top-down** approach.

Start handler: `(Observer, CompositeDisposable) -> Void`

When a depth 0 `SignalProducer` is started, the start handler would be invoked after the observer has been setup properly. The start handler may further start one or more depth 1 `SignalProducer`s. The process is recursively repeated until a depth N `SignalProducer` no longer starts any depth N+1 `SignalProducer`.

For each `SignalProducer`, a first level of `Signal` is created so that it provides a hook for asynchronous interruption. The event emitter of this `Signal` would then be passed into the start handler. After that, depending on the start handler, more levels of `Signal` might be created, and it could be part of the aforementioned recursive `SignalProducer` starting process.

Operator lifting, for example, would create another level of `Signal` by applying a `Signal` operator. So if a depth N `SignalProducer` is a lifted producer and it is preceded by a `Signal` graph of height M, the `Signal` graph would be of height M+2 afterwards.

#### Lifted Producer
The graph is constructed with a **bottom-up** approach.

Start handler: `() -> (signal: Signal, start: () -> Void, interrupter: Disposable)`

Let's first assume only the leaves would be interruptible producers.

When a depth 0 `SignalProducer` is started, a depth 0 `Signal` is created to provide a hook for asynchronous interruption.

The start handler would then be invoked after the observer has been setup properly. It would fabricate a `Signal`, a start action and an interrupter. Events of the returned `Signal` would be forwarded to the depth 0 `Signal`, and the interrupter would be placed into the composite disposable of the depth 0 `SignalProducer`. Finally, the start action would be invoked.

The fabricated `Signal` is the resulting `Signal` of the `Signal` transformations applied by the depth 0 `SignalProducer` on the root node of the `Signal` graph of depth 1 to depth N `SignalProducer`.

Similarly, the fabricated start action is an aggregate of all start actions of depth 0 to depth N `SignalProducer`, but would still be invoked in a top-bottom order. It is more or less the same for the fabricated interrupter, except that **the interruptibility of the intermediate `SignalProducer` is irrelevant, as long as the interruption is guaranteed to be propagated to the leaves of the `SignalProducer` graph**.

Now, as the start handler now breaks apart the creation of `Signal` graph , the start action and the interruption into three separate entity, a start handler can decide whther or not it needs the ability to be individually interrupted for its own purpose, and just pass through the start action and the interrupter of the next level.

Generally speaking, lifted `Signal` operator do not need the individual interruptibility, since `Signal` cannot be interrupted, and it can rely on the `interrupted` event from the leaves of the `Signal` graph to clean up resources.

Therefore, when lifting a `Signal` operator - unary or binary - on one or more depth N `SignalProducer`s, it would create a depth N-1 `SignalProducer` (downstream) with a start handler that applies the `Signal` operator directly to the `Signal`s fabricated by the start handlers of the depth N `SignalProducer`s (upstreams).

In other words, it would be the leaves of the `SignalProducer` graph who first create their `Signal`s and have it returned to its upper level. The process repeats recursively until it reaches the root node. That is why it is a bottom-up approach.