@andersio 2016-12-07
> I think we need to try to clarify/simplify the code before merging. 

I do hope to figure out a way to unify both "modes" to burn down the giant switch in binary `lift`. But I have run out of ideas, and got my hands busy on other stuff. 🐝

Feel free to criticise and play with the branch for a better way out.
 

/cc @ReactiveCocoa/reactivecocoa