@andersio 2016-12-07
Heh, it turned out to be quite straightforward. 🙈