@andersio 2017-04-14
Updated and polished.