@andersio 2017-05-01
Note that this addressed #227, since now downstreams are interrupted by upstreams. IOW `interrupted` would now respect the async operators applied.